Rename IgnoreModules -> UninstallModules
RELNOTES=N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=304644233
