+++ java/dagger/hilt/android/processor/BUILD
@@ -25,9 +25,9 @@ java_library(
     visibility = ["//visibility:private"],
     exports = [
         "//java/dagger/hilt/android/processor/internal/androidentrypoint:processor_lib",
-        "//java/dagger/hilt/android/processor/internal/ignoremodules:processor_lib",
         "//java/dagger/hilt/android/processor/internal/testing:internal_test_root_processor_lib",
         "//java/dagger/hilt/android/processor/internal/testroot:test_root_processor",
+        "//java/dagger/hilt/android/processor/internal/uninstallmodules:processor_lib",
         "//java/dagger/hilt/processor/internal/aggregateddeps:processor_lib",
         "//java/dagger/hilt/processor/internal/definecomponent:processor_lib",
         "//java/dagger/hilt/processor/internal/generatesrootinput:processor_lib",
@@ -56,12 +56,12 @@ gen_maven_artifact(
         "//java/dagger/hilt/android/processor/internal/androidentrypoint:compiler_options",
         "//java/dagger/hilt/android/processor/internal/androidentrypoint:metadata",
         "//java/dagger/hilt/android/processor/internal/androidentrypoint:processor_lib",
-        "//java/dagger/hilt/android/processor/internal/ignoremodules:processor_lib",
         "//java/dagger/hilt/android/processor/internal/testing:internal_test_root_metadata",
         "//java/dagger/hilt/android/processor/internal/testing:internal_test_root_processor_lib",
         "//java/dagger/hilt/android/processor/internal/testing:test_application_entry_point_generator",
         "//java/dagger/hilt/android/processor/internal/testing:test_application_generator",
         "//java/dagger/hilt/android/processor/internal/testroot:test_root_processor_lib",
+        "//java/dagger/hilt/android/processor/internal/uninstallmodules:processor_lib",
         "//java/dagger/hilt/codegen:originating_element",
         "//java/dagger/hilt/processor/internal:base_processor",
         "//java/dagger/hilt/processor/internal:classnames",
